# src/backend/docs/swagger.yaml
basePath: /api
definitions:
  internal_handlers.ChatRequest:
    properties:
      message:
        type: string
      session_id:
        type: string
    required:
    - message
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for AI Category Manager
  title: AI Category Manager API
  version: "1.0"
paths:
  /actions:
    get:
      description: Retrieve a list of actions, optionally filtered by status
      parameters:
      - description: Action status (e.g., pending, approved)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all actions
      tags:
      - actions
  /actions/{id}/approve:
    post:
      description: Approve or reject a specific action by ID
      parameters:
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update action status
      tags:
      - actions
  /actions/{id}/reject:
    post:
      description: Approve or reject a specific action by ID
      parameters:
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update action status
      tags:
      - actions
  /actions/generate:
    post:
      description: Use the AI Recommender agent to generate new action suggestions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate new actions
      tags:
      - actions
  /chat:
    post:
      consumes:
      - application/json
      description: Process user message through Supervisor Agent and stream response
        via SSE
      parameters:
      - description: Chat Message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ChatRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE Stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle chat messages
      tags:
      - chat
  /chat/sessions:
    get:
      description: Retrieve recent chat sessions with their first message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get chat sessions
      tags:
      - chat
  /chat/sessions/{id}/messages:
    get:
      description: Retrieve message history for a specific chat session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get chat messages for session
      tags:
      - chat
...
swagger: "2.0"